split messages into their own files
