+++ com.unity.ml-agents/Runtime/SideChannels/IncomingMessage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8043cec65aeb4ec09db1d25ad694328
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
