+++ com.unity.ml-agents/Runtime/SideChannels/OutgoingMessage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a007135a9a1e49849eb2d295f4c3879
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
