+++ ml-agents-envs/mlagents_envs/side_channel/__init__.py
@@ -1,0 +1,4 @@
+from mlagents_envs.side_channel.incoming_message import IncomingMessage  # noqa
+from mlagents_envs.side_channel.outgoing_message import OutgoingMessage  # noqa
+
+from mlagents_envs.side_channel.side_channel import SideChannel  # noqa
